refactor(css): C3 R3 design re-gate — move dppx precision-alignment to elidex_css evaluator
The cumulative /elidex-review (5-agent) on the R-loop delta flagged 1 IMP (Axis 1
Layering + Axis 3): R3's media-comparison precision-alignment was placed in the boa
producer (`bridge/media.rs` fed `resolution_dppx: f64::from(device_pixel_ratio() as
f32)`), but it encodes engine-independent CSS-comparison knowledge (the device dppx
must share the cssparser-f32 query precision for the exact `RangeValue::Dppx`
compare). Placing it at one producer (a) diverged from the VM producer
(`vm/host/media_query.rs` feeds f64, no narrowing) — so the VM, the S5-FLIP survivor,
would re-introduce the `(resolution: 1.2dppx)` mismatch — and (b) left the
`MediaEnvironment { resolution_dppx }` contract with an undocumented "must be
f32-quantized" precondition.

Fix: own the alignment in the engine-independent evaluator —
`elidex_css::media::eval::range_feature_value` narrows the device `resolution_dppx`
to f32 for the Resolution feature, so every producer (boa bridge + VM host) feeds the
lossless f64 and the evaluator aligns it to the query precision once. boa
`media_environment` reverts to feeding the f64 directly. `Converted` (dpi/dpcm) is
unaffected (its `VALUE_REL_EPS` tolerance absorbs the ≤6e-8 f32 narrowing); integer
dppx (exact in f32) is unchanged. elidex-css 462 tests + the fractional_dppx
regression test pass.

+++ crates/css/elidex-css/src/media/eval.rs
@@ -152,7 +152,18 @@ fn range_feature_value(name: RangeFeature, env: &MediaEnvironment) -> f64 {
         // §4.3: width / height. A degenerate zero height yields ±inf or NaN
         // (f64 division), which `compare` handles correctly — not a collapse to 0.
         RangeFeature::AspectRatio => env.viewport_width / env.viewport_height,
-        RangeFeature::Resolution => env.resolution_dppx,
+        // Narrow the device resolution to the query's precision before comparison. A
+        // queried `<resolution>` is tokenized by cssparser as `f32` (`parse.rs` →
+        // `Dppx(f64::from(f32))`) and an exact `RangeValue::Dppx` compares without
+        // tolerance, so a device dppx carried at full `f64` (e.g. a real 1.2× display =
+        // 1.199…956) would never equal the f32-sourced query 1.2 (1.200…0476) —
+        // `(resolution: 1.2dppx)` would spuriously fail. The comparison precision floor
+        // is cssparser's `f32`, so align the device to it. Engine-independent (every
+        // `MediaEnvironment` producer — boa bridge, VM host — feeds the lossless `f64`;
+        // the evaluator owns this alignment, not each producer). `Converted` (dpi/dpcm)
+        // already compares with `VALUE_REL_EPS`, which absorbs the ≤6e-8 f32 narrowing.
+        #[allow(clippy::cast_possible_truncation)]
+        RangeFeature::Resolution => f64::from(env.resolution_dppx as f32),
         RangeFeature::Color => f64::from(env.color_bits),
     }
 }
